--- MedscribeUI/src/components/Auth/authLayout.tsx
@@ -1,13 +1,0 @@
-import { Title } from '@mantine/core';
-
-function AuthLayout() {
-  return (
-    <div className="h-screen w-full flex">
-      <Title order={2} className="text-black text-center mb-6">
-        Auth Page... Coming Soon!
-      </Title>
-    </div>
-  );
-}
-
-export default AuthLayout;
